move the whole folder to under /src
