+++ frontend/jest.config.js
@@ -6,14 +6,6 @@ module.exports = {
   transform: {
     "^.+\\.[t|j]sx?$": "babel-jest",
   },
-  // transformIgnorePatterns: [
-  //     '/node_modules/(?!(axios)/)', // Tell Jest to transform axios
-  // ],
-
-  // transform: {
-  //     "\\.tsx?$": "ts-jest",
-  //     "\\.jsx?$": "babel-jest",
-  // },
   transformIgnorePatterns: ["/node_modules/(?!(axios)).+\\.js$"],
   testEnvironment: "jsdom", // Set the environment for testing React components
   moduleNameMapper: {
